examples/bitcoin-trading-app/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.34.0
python-telegram-bot==20.7
ccxt==4.1.13
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
plotly==5.18.0
websockets==12.0
